Plr trades definitions

+++ apisrc/EfhTestFuncs.c
@@ -412,7 +412,7 @@ void* onQuote(const EfhQuoteMsg* msg, EfhSecUserData secData, EfhRunUserData use
   int64_t priceScaleFactor = exch == EkaSource::kCME_SBE ? CME_DEFAULT_DISPLAY_PRICE_SCALE : DEFAULT_DISPLAY_PRICE_SCALE;
 #else
   int secIdx                  = (int)secData;
-  if (secIdx < 0 || secIdx >= gr->security.size())
+  if (secIdx < 0 || secIdx >= (int)gr->security.size())
     on_error("Bad secIdx = %d, gr->security.size() = %d",
 	     secIdx, (int)gr->security.size());
